Move board vec indexer to separate file
